… babyai/utils/agent.py
… scripts/eval_bot.py
… scripts/gui.py
… scripts/train_intelligent_expert.py