@maximecb , I went ahead and rewrote the bot. Highlights:

- 834 lines of code instead of 1080
- 1.5 times faster 
- still works flawlessly on 1M demos
- easier to understand and extend
- still can advise

For the record, here are some evaluation results:

**New bot, advising on 200 levels of BossLevel:**
BossLevel: 100.0%, r=0.930, s=109.86
total time: 46.5s
total episode_steps: 21973
total bfs: 39456
total bfs steps: 3256586

**Old bot, same**
BossLevel: 100.0%, r=0.930, s=110.75
total time: 60.7s
total episode_steps: 22149
total bfs: 74908
total bfs steps: 6096764

**New bot, acting on 200 levels of BossLevel:**
BossLevel: 100.0%, r=0.945, s=86.39
total time: 29.6s
total episode_steps: 17277
total bfs: 32170
total bfs steps: 2813974

**Old bot, same**
BossLevel: 100.0%, r=0.945, s=86.08
total time: 44.5s
total episode_steps: 17217
total bfs: 60084
total bfs steps: 5184787  

I think #30 can be abandoned in favor of this code. 

Pretty much all the docstrings are yet to be rewritten, but I guess you can get the idea of what's going on even without them, cause the code is now much, much simpler, IMHO.