@rizar 2019-01-28
It's more documented now. I considered implemented #29 , but I got feeling that it can be somewhat tricky and should be postponed for the brighter future to come.